Introduce FaLayersDirective and refactor FaLayersComponent to use host directives
