+++ src/lib/layers/layers.directive.ts
@@ -1,0 +1,35 @@
+import { Directive, inject, OnInit, input, computed, DOCUMENT } from '@angular/core';
+import { SizeProp } from '@fortawesome/fontawesome-svg-core';
+import { FaConfig } from '../config';
+import { ensureCss } from '../shared/utils/css';
+
+@Directive({
+  selector: '[faLayers]',
+  host: {
+    '[class]': 'classes()',
+  },
+})
+export class FaLayersDirective implements OnInit {
+  readonly size = input<SizeProp>();
+  readonly fixedWidth = input<boolean>();
+  readonly faFw = computed(() => {
+    const fixedWidth = this.fixedWidth();
+    return typeof fixedWidth === 'boolean' ? fixedWidth : this.config.fixedWidth;
+  });
+  readonly classes = computed(() => {
+    const sizeValue = this.size();
+    const sizeClass = sizeValue ? { [`fa-${sizeValue}`]: true } : {};
+    return {
+      ...sizeClass,
+      'fa-fw': this.faFw(),
+      'fa-layers': true,
+    };
+  });
+
+  private readonly document = inject(DOCUMENT);
+  private readonly config = inject(FaConfig);
+
+  ngOnInit() {
+    ensureCss(this.document, this.config);
+  }
+}
